+++ dev/20190415/k2works/fizzbuzz/src/main/java/org/hiroshima_arc/domain/model/fizzbuzz/FizzBuzzEntity.java
@@ -1,0 +1,20 @@
+package org.hiroshima_arc.domain.model.fizzbuzz;
+
+public class FizzBuzzEntity {
+    private Integer _number;
+    private String _value;
+
+    public FizzBuzzEntity(FizzBuzzValue value) {
+        _number = value.getNumber();
+        _value = value.getValue();
+    }
+
+    public Integer getNumber() {
+        return _number;
+    }
+
+    public String getValue() {
+        return _value;
+    }
+
+}
